+++ src/main/java/com/backend/digitalhouse/integrador/clinicaodontologica/dao/dto/entrada/modificacion/PacienteModificacionEntradaDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class PacienteModificacionEntradaDto {
     @NotNull
-    private int id;
+    private Long id;
     @NotNull
     private String nombre;
     @NotNull
@@ -21,7 +21,8 @@ public class PacienteModificacionEntradaDto {
     private DomicilioModificacionEntradaDto domicilio;
     public PacienteModificacionEntradaDto() {
     }
-    public PacienteModificacionEntradaDto(int id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioModificacionEntradaDto domicilio) {
+
+    public PacienteModificacionEntradaDto(Long id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioModificacionEntradaDto domicilio) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -30,11 +31,11 @@ public class PacienteModificacionEntradaDto {
         this.domicilio = domicilio;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
